Moved RGO randomization to new parser.
