+++ EU4toV2/CMakeLists.txt
@@ -18,6 +18,7 @@ file(GLOB MAIN_SOURCES "${PROJECT_SOURCE_DIR}/*.cpp")
 file(GLOB VIC2WORLD_SOURCES "${PROJECT_SOURCE_DIR}/V2World/*.cpp")
 file(GLOB V2FACTORY_SOURCES "${PROJECT_SOURCE_DIR}/V2World/Factory/*.cpp")
 file(GLOB V2LEADER_SOURCES "${PROJECT_SOURCE_DIR}/V2World/Leader/*.cpp")
+file(GLOB V2RGORANDOMIZATION_SOURCES "${PROJECT_SOURCE_DIR}/V2World/RGORandomization/*.cpp")
 file(GLOB HELPER_SOURCES "${PROJECT_SOURCE_DIR}/Helpers/*.cpp")
 file(GLOB MAPPER_SOURCES "${PROJECT_SOURCE_DIR}/Mappers/*.cpp")
 file(GLOB IDEAS_MAPPER_SOURCES "${PROJECT_SOURCE_DIR}/Mappers/Ideas/*.cpp")
@@ -52,7 +53,7 @@ set(Boost_USE_MULTITHREADED     OFF)
 set(Boost_USE_STATIC_RUNTIME    OFF)
 find_package(Boost)
 if(Boost_FOUND)
-  add_executable(EU4ToVic2 ${MAIN_SOURCES} ${VIC2WORLD_SOURCES} ${HELPER_SOURCES} ${MAPPER_SOURCES} ${IDEAS_MAPPER_SOURCES} ${V2LEADER_SOURCES} ${V2FACTORY_SOURCES} ${PROVINCE_MAPPER_SOURCES} ${EU4_WORLD_SOURCES} ${EU4_BUILDINGS_SOURCES} ${EU4_MODS_SOURCES} ${EU4_HISTORY_SOURCES} ${EU4_LEADER_SOURCES} ${EU4_MODIFIERS_SOURCES} ${EU4_PROVINCES_SOURCES} ${EU4_COUNTRY_SOURCES} ${EU4_ARMY_SOURCES} ${EU4_REGIONS_SOURCES} ${EU4_RELIGIONS_SOURCES} ${COMMON_SOURCES})
+  add_executable(EU4ToVic2 ${MAIN_SOURCES} ${VIC2WORLD_SOURCES} ${HELPER_SOURCES} ${MAPPER_SOURCES} ${IDEAS_MAPPER_SOURCES} ${V2RGORANDOMIZATION_SOURCES} ${V2LEADER_SOURCES} ${V2FACTORY_SOURCES} ${PROVINCE_MAPPER_SOURCES} ${EU4_WORLD_SOURCES} ${EU4_BUILDINGS_SOURCES} ${EU4_MODS_SOURCES} ${EU4_HISTORY_SOURCES} ${EU4_LEADER_SOURCES} ${EU4_MODIFIERS_SOURCES} ${EU4_PROVINCES_SOURCES} ${EU4_COUNTRY_SOURCES} ${EU4_ARMY_SOURCES} ${EU4_REGIONS_SOURCES} ${EU4_RELIGIONS_SOURCES} ${COMMON_SOURCES})
   add_custom_command(TARGET EU4ToVic2 POST_BUILD WORKING_DIRECTORY ${CMAKE_SOURCE_DIR} COMMAND chmod u+x Copy_Files.sh)
   add_custom_command(TARGET EU4ToVic2 POST_BUILD WORKING_DIRECTORY ${CMAKE_SOURCE_DIR} COMMAND ./Copy_Files.sh)
 endif()
