+++ EU4toV2/Copy_Files.bat
@@ -29,7 +29,7 @@ copy "Data_Files\port_blacklist.txt" "..\Release\EU4toVic2\port_blacklist.txt"
 copy "Data_Files\port_whitelist.txt" "..\Release\EU4toVic2\port_whitelist.txt"
 copy "Data_Files\minorityPops.txt" "..\Release\EU4toVic2\minorityPops.txt"
 copy "Data_Files\terrainData.txt" "..\Release\EU4toVic2\terrainData.txt"
-copy "Data_Files\shuffle_config.txt" "..\Release\EU4toVic2\shuffle_config.txt"
+copy "Data_Files\rgo_randomization.txt" "..\Release\EU4toVic2\rgo_randomization.txt"
 copy "Data_Files\FAQ.txt" "..\Release\EU4toVic2\FAQ.txt"
 copy "Data_Files\after_converting.txt" "..\Release\EU4toVic2\after_converting.txt"
 copy "Data_Files\Eu4ToVic2DefaultConfiguration.xml" "..\Release\Configuration\Eu4ToVic2DefaultConfiguration.xml"
