+++ EU4toV2/EU4ToV2.vcxproj
@@ -227,7 +227,6 @@
     <ClCompile Include="Source\EU4World\Religions\Religions.cpp" />
     <ClCompile Include="Source\EU4World\World.cpp" />
     <ClCompile Include="Source\FlagUtils.cpp" />
-    <ClCompile Include="Source\Helpers\RgoShuffle.cpp" />
     <ClCompile Include="Source\Helpers\TechValues.cpp" />
     <ClCompile Include="Source\main.cpp" />
     <ClCompile Include="Source\Mappers\AdjacencyMapper.cpp" />
@@ -275,6 +274,8 @@
     <ClCompile Include="Source\V2World\Leader\V2LeaderTrait.cpp" />
     <ClCompile Include="Source\V2World\Leader\V2LeaderTraitMapper.cpp" />
     <ClCompile Include="Source\V2World\Leader\V2LeaderTraitTypes.cpp" />
+    <ClCompile Include="Source\V2World\RGORandomization\Bucket.cpp" />
+    <ClCompile Include="Source\V2World\RGORandomization\BucketList.cpp" />
     <ClCompile Include="Source\V2World\StateMapper.cpp" />
     <ClCompile Include="Source\V2World\V2UncivReforms.cpp" />
     <ClCompile Include="Source\V2World\Vic2CultureUnion.cpp" />
@@ -358,7 +359,6 @@
     <ClInclude Include="Source\EU4World\Religions\Religions.h" />
     <ClInclude Include="Source\EU4World\World.h" />
     <ClInclude Include="Source\FlagUtils.h" />
-    <ClInclude Include="Source\Helpers\RgoShuffle.h" />
     <ClInclude Include="Source\Helpers\TechValues.h" />
     <ClInclude Include="Source\Mappers\AdjacencyMapper.h" />
     <ClInclude Include="Source\Mappers\CK2TitleMapper.h" />
@@ -405,6 +405,8 @@
     <ClInclude Include="Source\V2World\Leader\V2LeaderTrait.h" />
     <ClInclude Include="Source\V2World\Leader\V2LeaderTraitMapper.h" />
     <ClInclude Include="Source\V2World\Leader\V2LeaderTraitTypes.h" />
+    <ClInclude Include="Source\V2World\RGORandomization\Bucket.h" />
+    <ClInclude Include="Source\V2World\RGORandomization\BucketList.h" />
     <ClInclude Include="Source\V2World\StateMapper.h" />
     <ClInclude Include="Source\V2World\V2Army.h" />
     <ClInclude Include="Source\V2World\V2Country.h" />
