+++ EU4toV2/EU4ToV2.vcxproj.filters
@@ -372,6 +372,12 @@
     <ClCompile Include="Source\V2World\Leader\V2LeaderTraitMapper.cpp">
       <Filter>Vic2World\Leader</Filter>
     </ClCompile>
+    <ClCompile Include="Source\V2World\RGORandomization\Bucket.cpp">
+      <Filter>Vic2World\RGORandomization</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\V2World\RGORandomization\BucketList.cpp">
+      <Filter>Vic2World\RGORandomization</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Source\Configuration.h" />
@@ -747,6 +753,12 @@
     <ClInclude Include="Source\V2World\Leader\V2LeaderTraitMapper.h">
       <Filter>Vic2World\Leader</Filter>
     </ClInclude>
+    <ClInclude Include="Source\V2World\RGORandomization\Bucket.h">
+      <Filter>Vic2World\RGORandomization</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\V2World\RGORandomization\BucketList.h">
+      <Filter>Vic2World\RGORandomization</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="EU4World">
@@ -805,6 +817,9 @@
     </Filter>
     <Filter Include="EU4World\Leader">
       <UniqueIdentifier>{cad4ad12-a780-4d35-bce5-cfa1c1b23f89}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Vic2World\RGORandomization">
+      <UniqueIdentifier>{c2a72362-f9f1-46ba-9af1-ea88983f295c}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
 </Project>
